# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "Category": {
        "enum": [
          "BLOQUEO_TARJETA",
          "REPORTE_FRAUDE",
          "CONSULTA_GENERAL",
          "SOLICITUD_CREDITO",
          "BANCA_DIGITAL"
        ],
        "title": "Category",
        "type": "string"
      },
...
      "ConsultationLevel": {
        "enum": [
          "GENERAL",
          "PERSONALIZADA",
          "SENSIBLE"
        ],
        "title": "ConsultationLevel",
        "type": "string"
      },
...
      "ConversationMessageOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "role": {
            "$ref": "#/components/schemas/ConversationRole"
          },
          "text": {
            "title": "Text",
            "type": "string"
          }
        },
        "required": [
          "id",
          "role",
          "text",
          "created_at"
        ],
        "title": "ConversationMessageOut",
        "type": "object"
      },
      "ConversationRole": {
        "enum": [
          "CUSTOMER",
          "ASSISTANT"
        ],
        "title": "ConversationRole",
        "type": "string"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "IdentificationStatus": {
        "enum": [
          "ANONIMO",
          "PENDIENTE",
          "IDENTIFICADO",
          "FALLIDO",
          "VERIFICACION_MANUAL"
        ],
        "title": "IdentificationStatus",
        "type": "string"
      },
      "IdentifierRevealResponse": {
        "properties": {
          "identifier": {
            "title": "Identifier",
            "type": "string"
          },
          "reveal_seconds": {
            "default": 30,
            "title": "Reveal Seconds",
            "type": "integer"
          }
        },
        "required": [
          "identifier"
        ],
        "title": "IdentifierRevealResponse",
        "type": "object"
      },
...
      "Priority": {
        "enum": [
          "CRITICO",
          "ALTO",
          "MEDIO",
          "BAJO"
        ],
        "title": "Priority",
        "type": "string"
      },
      "ProtectedIdentity": {
        "properties": {
          "display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Display Name"
          },
          "masked_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Masked Identifier"
          },
          "reveal_available": {
            "default": false,
            "title": "Reveal Available",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/IdentificationStatus"
          }
        },
        "required": [
          "status"
        ],
        "title": "ProtectedIdentity",
        "type": "object"
      },
...
      "ResolutionOutcome": {
        "enum": [
          "RESUELTO",
          "DERIVADO",
          "PENDIENTE_DOCUMENTACION",
          "CLIENTE_DESISTIO",
          "NO_RESUELTO"
        ],
        "title": "ResolutionOutcome",
        "type": "string"
      },
...
      "TicketDetail": {
        "properties": {
          "category": {
            "$ref": "#/components/schemas/Category"
          },
          "client_display_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client Display Name"
          },
          "client_session_id": {
            "title": "Client Session Id",
            "type": "string"
          },
          "closed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Closed At"
          },
          "consultation_level": {
            "$ref": "#/components/schemas/ConsultationLevel"
          },
          "conversation": {
            "items": {
              "$ref": "#/components/schemas/ConversationMessageOut"
            },
            "title": "Conversation",
            "type": "array"
          },
          "estimated_wait_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Wait Minutes"
          },
          "events": {
            "items": {
              "$ref": "#/components/schemas/TraceEventOut"
            },
            "title": "Events",
            "type": "array"
          },
          "executive_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executive Name"
          },
          "executive_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Executive Title"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "identification_status": {
            "$ref": "#/components/schemas/IdentificationStatus"
          },
          "identity": {
            "$ref": "#/components/schemas/ProtectedIdentity"
          },
          "masked_identifier": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Masked Identifier"
          },
          "minutes_elapsed": {
            "title": "Minutes Elapsed",
            "type": "integer"
          },
          "number": {
            "title": "Number",
            "type": "string"
          },
          "preferential_attention": {
            "title": "Preferential Attention",
            "type": "boolean"
          },
          "priority": {
            "$ref": "#/components/schemas/Priority"
          },
          "resolution_note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "resolution_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          },
          "time_assigned": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Time Assigned"
          },
          "version": {
            "title": "Version",
            "type": "integer"
          },
          "wait_time_min": {
            "title": "Wait Time Min",
            "type": "integer"
          },
          "window_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Window Number"
          }
        },
        "required": [
          "id",
          "number",
          "category",
          "priority",
          "summary",
          "time_assigned",
          "minutes_elapsed",
          "executive_name",
          "executive_title",
          "window_number",
          "status",
          "client_session_id",
          "wait_time_min",
          "estimated_wait_minutes",
          "identification_status",
          "preferential_attention",
          "version",
          "consultation_level",
          "identity",
          "conversation",
          "events"
        ],
        "title": "TicketDetail",
        "type": "object"
      },
...
      "TicketPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/TicketListItem"
            },
            "title": "Items",
            "type": "array"
          },
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "status_counts": {
            "additionalProperties": {
              "type": "integer"
            },
            "propertyNames": {
              "$ref": "#/components/schemas/TicketStatus"
            },
            "title": "Status Counts",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "page",
          "page_size",
          "total",
          "status_counts"
        ],
        "title": "TicketPage",
        "type": "object"
      },
...
      "TicketStatus": {
        "enum": [
          "PENDIENTE",
          "EN_ATENCION",
          "CERRADO"
        ],
        "title": "TicketStatus",
        "type": "string"
      },
      "TicketStatusUpdate": {
        "properties": {
          "expected_version": {
            "minimum": 1.0,
            "title": "Expected Version",
            "type": "integer"
          },
          "resolution_note": {
            "anyOf": [
              {
                "maxLength": 1000,
                "minLength": 10,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolution Note"
          },
          "resolution_outcome": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ResolutionOutcome"
              },
              {
                "type": "null"
              }
            ]
          },
          "status": {
            "$ref": "#/components/schemas/TicketStatus"
          }
        },
        "required": [
          "status",
          "expected_version"
        ],
        "title": "TicketStatusUpdate",
        "type": "object"
      },
...
      "TraceEventOut": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "description": {
            "title": "Description",
            "type": "string"
          },
          "event_type": {
            "title": "Event Type",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "metadata": {
            "additionalProperties": true,
            "title": "Metadata",
            "type": "object"
          }
        },
        "required": [
          "id",
          "event_type",
          "description",
          "created_at"
        ],
        "title": "TraceEventOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "API del sistema de orquestacion de atencion bancaria",
    "title": "Sistema de Orquestacion Bancaria",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/executive/tickets": {
      "get": {
        "operationId": "executive_tickets_api_v1_executive_tickets_get",
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TicketStatus"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Category"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "priority",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/Priority"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Priority"
            }
          },
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 120,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "priority",
              "enum": [
                "priority",
                "oldest",
                "newest"
              ],
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Executive Tickets",
        "tags": [
          "Operacion ejecutiva"
        ]
      }
    },
...
    "/api/v1/tickets/{ticket_id}": {
      "get": {
        "operationId": "ticket_detail_api_v1_tickets__ticket_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Ticket Detail",
        "tags": [
          "Operacion ejecutiva"
        ]
      }
    },
    "/api/v1/tickets/{ticket_id}/identifier/reveal": {
      "post": {
        "operationId": "reveal_identifier_api_v1_tickets__ticket_id__identifier_reveal_post",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IdentifierRevealResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Reveal Identifier",
        "tags": [
          "Operacion ejecutiva"
        ]
      }
    },
    "/api/v1/tickets/{ticket_id}/status": {
      "patch": {
        "operationId": "update_ticket_status_api_v1_tickets__ticket_id__status_patch",
        "parameters": [
          {
            "in": "path",
            "name": "ticket_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Ticket Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TicketStatusUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TicketDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Update Ticket Status",
        "tags": [
          "Operacion ejecutiva"
        ]
      }
    }
  }
}